+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [2.12.8] - 2019-05-02
+### Added
+- NewConnector, NewSessionIniter
+
+## [2.12.7] - 2019-04-24
+### Changed
+- ODPI-C v3.1.4 (rowcount for PL/SQL block)
 
 ## [2.12.6] - 2019-04-12
 ### Added